chore(core): reorganize api client package (#109)
