+++ client/go/bot/option.go
@@ -3,7 +3,7 @@ package bot
 import (
 	grpc "google.golang.org/grpc"
 
-	"github.com/berty/berty/core/client"
+	"github.com/berty/berty/core/api/client"
 )
 
 type Option func(b *Bot) error
