+++ core/api/client/jsonclient/{{.File.Package}}.{{.Service.Name|lower}}.gen.go.tmpl
@@ -10,7 +10,7 @@ import (
 
         "go.uber.org/zap"
 
-	"github.com/berty/berty/core/client"
+	"github.com/berty/berty/core/api/client"
         "github.com/berty/berty/core/{{.File.Name|dir}}"
         {{- range .File.Dependency}}
           {{- $dependency := . | getProtoFile}}
